requirements.txt
# Sam3 Segmentation Service Dependencies

# FastAPI framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# HTTP client for downloading images
httpx>=0.26.0

# ML / Model inference
# SAM3SemanticPredictor was added in ultralytics 8.3.237 (PR #22897)
ultralytics>=8.3.237
timm>=1.0.0          # required by ultralytics SAM3 backbone
pillow>=10.1.0
numpy>=1.24.0

# CLIP — SAM3's text encoder requires the ultralytics fork of CLIP.
# The standard PyPI `clip` package causes "SimpleTokenizer not callable" errors.
git+https://github.com/ultralytics/CLIP.git

# PyTorch (CPU build - swap for CUDA build on GPU Railway instance)
torch>=2.1.0
torchvision>=0.16.0

# S3 / Railway bucket - model download
boto3>=1.34.0

# Database (shared PostgreSQL for authentication, same as gateway)
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Environment management
python-dotenv>=1.0.0
